# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AmexingWeb API",
    "version": "0.1.0",
    "description": "\n# AmexingWeb E-Commerce Platform API\n\nPCI DSS Level 1 compliant e-commerce platform with comprehensive security features.\n\n## Features\n- **Authentication**: Multi-provider OAuth 2.0 (Username/Password, Apple, Corporate)\n- **Security**: JWT Bearer tokens with secure HTTP-only cookies\n- **Authorization**: Role-based access control (RBAC) with 7 role levels\n- **Data Protection**: Encryption at rest and in transit\n- **Audit Logging**: Comprehensive audit trails for all operations\n\n## Security Standards\n- PCI DSS 4.0.1 Compliant\n- OWASP Top 10 Protection\n- Rate Limiting on all endpoints\n- CSRF Protection for state-changing operations\n- Security Headers (Helmet.js)\n\n## Authentication Flow\n1. Obtain JWT token via `/auth/login` or OAuth providers\n2. Include token in `Authorization: Bearer <token>` header\n3. Token lifetime: 8 hours (access), 7 days (refresh)\n4. Use `/auth/refresh` to renew tokens before expiration\n\n## Rate Limiting\n- Authentication endpoints: 50 requests / 15 minutes\n- API endpoints: 100 requests / 15 minutes\n- Write operations: 30 requests / 15 minutes\n- Password reset: 10 requests / 5 minutes\n\n## Support\nFor API support, contact the Amexing Development Team.\n      ",
    "contact": {
      "name": "Amexing Development Team",
      "email": "dev@amexing.com"
    },
    "license": {
      "name": "MIT",
...
    }
  },
  "servers": [
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "\nJWT Bearer token authentication.\n\n**Token Lifetime (PCI DSS Compliant):**\n- Access Token: 8 hours\n- Refresh Token: 7 days\n\n**How to obtain a token:**\n1. Use `POST /auth/login` with credentials\n2. Or authenticate via OAuth providers\n3. Token will be returned in response and set as HTTP-only cookie\n\n**Token includes:**\n- userId: User unique identifier\n- username: Username\n- email: User email\n- role: User role (superadmin, admin, client, etc.)\n- roleId: Role object ID\n- organizationId: Organization identifier\n- name: User display name\n\n**Security Features:**\n- Tokens are signed with HS256 algorithm\n- HTTP-only cookies prevent XSS attacks\n- Secure flag in production (HTTPS only)\n- SameSite=strict prevents CSRF\n          "
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "\nHTTP-only cookie containing JWT access token.\n\n**Security Features:**\n- HttpOnly: Cannot be accessed via JavaScript\n- Secure: Only transmitted over HTTPS in production\n- SameSite=strict: Prevents CSRF attacks\n- Max-Age: 8 hours\n\nThis is automatically set by the server upon successful authentication.\n          "
      }
    },
...
    "parameters": {
      "PageParameter": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "description": "Page number for pagination"
      },
...
    },
    "schemas": {
...
      "SystemStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "healthy",
              "degraded",
              "down"
            ],
            "example": "healthy"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-01T12:00:00.000Z"
          },
          "services": {
            "type": "object",
            "properties": {
              "database": {
                "type": "string",
                "enum": [
                  "connected",
                  "disconnected"
                ],
                "example": "connected"
              },
              "parseServer": {
                "type": "string",
                "enum": [
                  "running",
                  "stopped"
                ],
                "example": "running"
              }
            }
          }
        }
      },
...
      "NotificationsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Notifications retrieved successfully"
          },
          "data": {
            "type": "object",
            "properties": {
              "notifications": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Notification"
                }
              },
              "unreadCount": {
                "type": "integer",
                "description": "Number of unread notifications",
                "example": 5
              },
              "totalCount": {
                "type": "integer",
                "description": "Total number of notifications",
                "example": 50
              }
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/status": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Get API status",
        "description": "Check API health and service availability.\n\n**Public Endpoint** - No authentication required\n**Rate Limited:** 100 requests per 15 minutes\n\n**Health Checks:**\n- API server status\n- Database connectivity\n- Parse Server status\n",
        "responses": {
          "200": {
            "description": "API is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatus"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatus"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get user notifications",
        "description": "Retrieve notifications for authenticated user.\n\n**Access:** Requires 'notifications.read' permission\n**Rate Limited:** 100 requests per 15 minutes\n",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/notifications/{notificationId}/read": {
      "patch": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark notification as read",
        "description": "Mark a specific notification as read.\n\n**Access:** Requires 'notifications.update' permission\n",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "notificationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Notification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marked as read"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/api/notifications/mark-all-read": {
      "patch": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark all notifications as read",
        "description": "Mark all user notifications as read.\n\n**Access:** Requires 'notifications.update' permission\n",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All notifications marked as read"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/user/profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Get current user profile",
        "description": "Retrieve authenticated user's profile information.\n\n**Access:** Requires 'profile.read' permission (all authenticated users)\n**Rate Limited:** 100 requests per 15 minutes\n",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "put": {
        "tags": [
          "Profile"
        ],
        "summary": "Update current user profile",
        "description": "Update authenticated user's profile information.\n\n**Updatable Fields:**\n- firstName, lastName\n- Email (requires re-verification)\n- Phone\n- Preferences\n\n**Access:** Requires 'profile.update' permission\n**Rate Limited:** 100 requests per 15 minutes\n",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": [
          "User Management"
        ],
        "summary": "Get list of users with filtering and pagination",
        "description": "Retrieve paginated list of users with advanced filtering options.\n\n**Access Control:**\n- Requires authentication (JWT token)\n- Requires 'users.list' permission\n- SuperAdmin/Admin: See all users\n- DepartmentManager: See users in their department\n- Others: Limited access based on role\n\n**Features:**\n- Pagination (default: 25 items, max: 100)\n- Filter by role, status, client, department\n- Sort by any field\n- Search functionality\n- Rate limited: 100 requests per 15 minutes\n\n**PCI DSS:**\n- Audit logged per requirement 10.2.1\n- Role-based data filtering\n",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParameter"
          },
          {
            "$ref": "#/components/parameters/LimitParameter"
          },
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "superadmin",
                "admin",
                "client",
                "department_manager",
                "employee",
                "employee_amexing",
                "driver",
                "guest"
              ]
            },
            "description": "Filter by user role"
          },
          {
            "in": "query",
            "name": "active",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by active status"
          },
          {
            "in": "query",
            "name": "emailVerified",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by email verification status"
          },
          {
            "in": "query",
            "name": "clientId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by client (admin only)"
          },
          {
            "in": "query",
            "name": "departmentId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by department"
          },
          {
            "$ref": "#/components/parameters/SortFieldParameter"
          },
          {
            "$ref": "#/components/parameters/SortDirectionParameter"
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "User Management"
        ],
        "summary": "Create new user",
        "description": "Create a new user account with role assignment.\n\n**Access:** Requires 'users.create' permission\n**Password:** Auto-generated if not provided\n**Rate Limited:** 30 requests per 15 minutes\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
...
  }
}